Update project to target .NET 9.0 and revise dependencies
The project file `MauiBlazorBridge.csproj` has been updated to target .NET 9.0, including changes to the `<TargetFrameworks>` and `<SupportedOSPlatformVersion>` elements for various platforms. The `Microsoft.AspNetCore.Components.Web` package version has been upgraded to 9.0.3. The project version is now set to `1.0.0`, and the summary and description have been updated to include "Internet" in the context of platform detection.